profiles/requirements.txt
Jinja2==2.6
PyYAML==3.10
Werkzeug==0.8.3
decorator==3.4.0
gunicorn==0.17.4
horsemeat==2.12.9
ipython==0.13.2
psycopg2==2.5
superlance==0.7
supervisor==3.0b1
wsgiref==0.1.2
soloprocess==0.0.4
s3cmd==1.5.2
itsdangerous
logtopg>=0.1.1
colorlog==2.4.0
pillow==2.9.0
pyrax==1.9.7
